move is_function_invoke_mode_response_stream into lambda invoke executors and add/update string constants
